Calculate distances using distance maps
Adds support for reading distance map JSONs and calculating distances between
sequences using a given map and the sequences in a given tree. This support
replaces the original mask-based definitions for calculating distances and
provides a more generic way to calculate distances between sequences.

This commit only adds support for comparisons between all nodes and the tree's
root. Future commits will add support for comparison to the last ancestor in a
previous season or pairwise comparisons to all samples in previous seasons.